fix: Refactored the front-end code to make vue support multiple pages
