+++ witstalk/package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "axios": "^0.24.0",
+    "js-md5": "^0.7.3",
     "vue": "^2.5.2",
     "vue-router": "^3.0.1"
   },
