--- witstalk/src/App.vue
@@ -1,16 +1,0 @@
-<template>
-  <div id="app">
-    <index msg="Welcome to Your Vue.js App"/>
-  </div>
-</template>
-
-<script>
-import index from './components/index.vue'
-
-export default {
-  name: 'app',
-  components: {
-    index
-  }
-}
-</script>
