--- witstalk/src/components/chat.vue
@@ -1,8 +1,0 @@
-<template>
-</template>
-
-<script>
-</script>
-
-<style scoped>
-</style>
