+++ witstalk/src/pages/Login/Login.js
@@ -1,0 +1,15 @@
+// The Vue build version to load with the `import` command
+// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
+import Vue from 'vue'
+import App from './Login.vue'
+import router from './router/index'
+
+Vue.config.productionTip = false
+
+/* eslint-disable no-new */
+new Vue({
+  el: '#login',
+  router,
+  components: { App },
+  template: '<App/>'
+})
