+++ witstalk/src/pages/index/index.js
@@ -1,0 +1,15 @@
+// The Vue build version to load with the `import` command
+// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
+import Vue from 'vue'
+import App from './index.vue'
+import router from '../Login/router/index'
+
+Vue.config.productionTip = false
+
+/* eslint-disable no-new */
+new Vue({
+  el: '#index',
+  router,
+  components: { App },
+  template: '<App/>'
+})
