+++ witstalk/src/pages/index/index.vue
@@ -1,0 +1,10 @@
+<template>
+  <div>正在跳转</div>
+</template>
+
+<script>
+</script>
+
+<style scoped>
+
+</style>
